Improvements and PNG physical size handling.
Animated PNG-specific chunks (acTL, fcTL, fdAT) will now be ommited if the PNG only has one frame and is not an Animated PNG.

Fixed an issue with single-frame WEBPs not working.
(Although a single-frame WEBP is not an Animated WEBP and it may seem like that does not matter for this, there may be a purpose for using this to process non-Animated WEBP if features like metadata insertion get added later.)

PNG now supports the 'ppi' (pixels per inch) and 'ppm'(pixels per metre) paramter. If one of those are set, it will create a pHYs chunk inside the image so that printers and other hardware/software can print, display, or handle it at the correct real-life physical size.