Many improvements for more optimal encoding.
The animationStyle parameter is now removed and will be ignored if set. The encoder will automatically detect and use optimal frame disposal and blending modes on a frame-by-frame basis. In many animations, some parts can recycle most of the pixels and other parts of the animations need large parts updated with source blending to change areas to a transparency.

Byte filtering was added for additional size reduction. (Currently Sub and Up mode only.)

24-bit mode now has logic to select an unused color that pixels will not quantize to as the reserved transparent pixel color.

Various improvements.