Adds z-index channel splitting, merge-to-wire-channel floods, and visual improvements. Improves flood fill logic.
The wire channel can now be drawn on a z-index higher than the fill channel allowing the wire strokes to show through other bitmaps that cover the fill. The colors that should show over in this style need to have the zw property set to true. The bitmap image will need to be configured to have a split z-index if that is the desired behavior. Fill areas can now be merged to the wire channel with a flood tool. Various visual improvements were made. Flood fill logic was improved to better fill into tight corners. The wire flood can now directly fill in empty areas between wires to make a thick style stroke shape.