Adds ZIP metadata file arc.json, a proposed standard
Adds arc.json to the ZIP file root, which shall now be proposed as a standard for type validation in ZIP-based formats where a custom Magic Number is no an option. Also addresses issues like ensuring correct created date if manually unpacked and repacked, and giving specific details about how the file can be opened, which is useful for non-mainstream types. arc.json is now draft 1.0, read more about it in the GraFlicUtil.js documentaiton.