Fixed invalid ZIP formatting, improved z.json archive metadata.
Fixed invalid ZIP formatting in GraFlicArchive. It was interoperable with some things, but not others.
Tweaked z.json archive metadata/type validator for ZIP-based files.
Will now fix the common packaging mistake by looking for the z.json root location.
(Due to the wonky, unintuitive way that packaging is typically handled, users will often right-click the folder and archive rather than select and archive folder contents, resulting in files being buried in a redundant folder. If they select the files and do it the right way, then the archive can end up in the same folder as the contained files, which is also awkward...)
Other improvements to the ZIP formatting were made. GraFlicArchive will now use MS-DOS FAT as the Host OS since that seems to be the most widely supported and interoperable for external attributes. *NIX systems seem to do OK interpreting some of these attributes like Read-Only.
readInt32 was added and readUint32 was fixed to not overflow to negative on exceptionally large numbers.