Stops Touch and Pointer Events from colliding if a device has touch and pen/stylus.
In devices that have touch screens and a stylus, the Touch and Pointer events could collide, resulting in garbled strokes. Touch Events will be used when Pointer Events are not supported, but if Pointer Events are supported Touch Events will be disabled and it will let the Pointer Events handle stylus/mouse/touch. Force property may be implemented later for devices that support force as an alternative to the pressure property.